Move networks file from deposit to eth2 service
