+++ src/renderer/services/eth2/networks.ts
