+++ test/unit/eth2/client/prysm/beacon.spec.ts
@@ -1,5 +1,5 @@
 import {PrysmBeaconApiClient, PrysmBeaconRoutes} from "../../../../../src/renderer/services/eth2/client/prysm/beacon";
-import {networks} from "../../../../../src/renderer/services/deposit/networks";
+import {networks} from "../../../../../src/renderer/services/eth2/networks";
 import {SupportedNetworks} from "../../../../../src/renderer/services/docker/chain";
 import axios from "axios";
 import MockAxiosAdapter from "@nodefactory/axios-mock-adapter";
